requirements.txt
python-telegram-bot==20.7
pytest==7.4.3
pytest-asyncio==0.21.1
